refactor(settings): extract WeChat login modals and fields to own module
BotWeChatFields, WeChatScanLoginModal, and WechatQrLoginModal (the
self-contained WeChat scan/QR login flow, ~370 lines) lived inside
bot-chat-settings-page.tsx. Move them to bot-wechat-login.tsx and import
them back, dropping the now-unused X / DialogRoot / DialogContent /
WechatBridgeQrCodeResult imports from the host page.

bot-wechat-login.tsx is registered in SETTINGS_SOURCE_REPO_PATHS right
after bot-chat-settings-page.tsx so the contract source mirror still
contains the WeChat modal code.
